backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
google-generativeai==0.3.2
chromadb==0.4.22
langchain==0.1.0
langchain-community==0.0.10
pypdf2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
pydub==0.25.1
Werkzeug==3.0.1
requests==2.31.0
